REWORKED from ground up.

+++ src/main/java/lt/inventi/messaging/domain/IdContainer.java
@@ -3,17 +3,17 @@ package lt.inventi.messaging.domain;
 import java.math.BigInteger;
 
 public class IdContainer {
-    private BigInteger id;
+    private Long id;
 
-    public IdContainer(BigInteger id) {
+    public IdContainer(Long id) {
         this.id = id;
     }
 
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
